HW-5 BaseNaverSearchActivity를 만들어서 겹치는 코드 처리
