+++ bean628/MyProject/app/build.gradle
@@ -1,6 +1,7 @@
 apply plugin: 'com.android.application'
 apply plugin: 'kotlin-android'
 apply plugin: 'kotlin-kapt'
+apply plugin: 'kotlin-android-extensions'
 
 android {
     compileSdkVersion 28
@@ -27,7 +28,6 @@ android {
     dataBinding {
         enabled = true
     }
-
 
 
 }
